website/requirements.txt
# Core data pipeline
dlt[postgres]>=0.4.0
pandas>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Dashboard and visualization
streamlit>=1.28.0
plotly>=5.15.0

# AI and LangChain integration
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
openai>=1.0.0

# Environment management
python-dotenv>=1.0.0

# Additional utilities
pydantic>=2.0.0